Add ShadowJar to build executables
With the previous build.gradle overhaul, the configuration for jar was removed.
This was because the gradle jar command would build the application without embedding
the dependencies, which is vital to the run the application on it's own.

As such, the ShadowJar plugin has been added to the build script. Using the
plugin, a fat jar file (src + dependencies) can be generated, which can be
run without the need for downloading in dependencies.